requirements.txt
Flask==3.1.0
discord.py==2.4.0
requests==2.31.0
python-dotenv==1.0.1
PyGithub==2.5.0